made some minor fixes / layout changes
