+++ Makefile
@@ -1,9 +1,9 @@
 CC      = gcc
-CFLAGS  = -Wall -g -O3 -Iinclude
+CFLAGS  = -Wall -g -O3 -Iinclude -Ilib/include
 LIBS    = -lm -ldl -lglfw -lGL -lyaml
 
 # break this out into individual object files at some point
-SRC     = src/main.c src/physics/gravity.c src/glad.c src/math/math_funcs.c src/utils/constants.c src/utils/utils.c src/math/vector2.c src/physics/cr3bp.c
+SRC     = src/main.c src/physics/gravity.c lib/glad.c src/math/math_funcs.c src/utils/constants.c src/utils/utils.c src/math/vector2.c src/physics/cr3bp.c src/utils/config_parser.c
 OUTDIR  = out
 TARGET  = $(OUTDIR)/sim
 
@@ -19,4 +19,4 @@ clean:
 	rm -f $(OUTDIR)/*
 
 check: $(TARGET)
-	valgrind --leak-check=full ./$(TARGET)
+	valgrind --leak-check=full ./$(TARGET) -m n-body -n 4 -t 5000
