+++ include/utils/config_parser.h
@@ -1,0 +1,8 @@
+#ifndef PARSER_H
+#define PARSER_H
+
+#include "utils/structures.h"
+
+void parse_config_file(two_d_body* bodies_array[], int NUM_BODIES);
+
+#endif
